Patch tilegx v8 support.
 the node.js is cloned from official git on github, the v8 version
is different from our current v8, so we need to backport or upstream
our v8 code.

+++ deps/v8/src/objects.cc
@@ -8729,9 +8729,9 @@ static inline bool CompareRawStringContents(const Char* const a,
   // If this architecture isn't comfortable reading unaligned ints
   // then we have to check that the strings are aligned before
   // comparing them blockwise.
-  const int kAlignmentMask = sizeof(uint32_t) - 1;  // NOLINT
-  uint32_t pa_addr = reinterpret_cast<uint32_t>(a);
-  uint32_t pb_addr = reinterpret_cast<uint32_t>(b);
+  const uint64_t kAlignmentMask = sizeof(uint64_t) - 1;  // NOLINT
+  uint64_t pa_addr = reinterpret_cast<uint64_t>(a);
+  uint64_t pb_addr = reinterpret_cast<uint64_t>(b);
   if (((pa_addr & kAlignmentMask) | (pb_addr & kAlignmentMask)) == 0) {
 #endif
     const int kStepSize = sizeof(int) / sizeof(Char);  // NOLINT
